Move Math.hpp
