+++ Core/include/Voxymore/Core/YamlHelper.hpp
@@ -4,9 +4,9 @@
 
 #pragma once
 
-#include <yaml-cpp/yaml.h>
-#include "Voxymore/Core/Math.hpp"
+#include "Voxymore/Math/Math.hpp"
 #include <filesystem>
+#include <yaml-cpp/yaml.h>
 
 #define KEY(x) YAML::Key << x
 #define VAL(x) YAML::Value << x
