+++ Core/include/Voxymore/Renderer/Camera.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 #include "Voxymore/Core/Core.hpp"
-#include "Voxymore/Core/Math.hpp"
+#include "Voxymore/Math/Math.hpp"
 
 namespace Voxymore::Core {
 	class Camera {
