+++ Core/include/Voxymore/Renderer/Light.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include <Voxymore/Core/Math.hpp>
+#include "Voxymore/Math/Math.hpp"
 
 namespace Voxymore::Core
 {
