+++ Core/include/Voxymore/Renderer/Model.hpp
@@ -4,16 +4,16 @@
 
 #pragma once
 
-#include "Voxymore/Core/Macros.hpp"
-#include "Voxymore/Core/Logger.hpp"
-#include "Voxymore/Core/SmartPointers.hpp"
 #include "Voxymore/Core/FileSystem.hpp"
-#include "Voxymore/Core/Math.hpp"
+#include "Voxymore/Core/Logger.hpp"
+#include "Voxymore/Core/Macros.hpp"
+#include "Voxymore/Core/SmartPointers.hpp"
+#include "Voxymore/Math/Math.hpp"
 #include "Voxymore/Renderer/Buffer.hpp"
-#include "Voxymore/Renderer/VertexArray.hpp"
-#include "Voxymore/Renderer/UniformBuffer.hpp"
-#include "Voxymore/Renderer/Shader.hpp"
 #include "Voxymore/Renderer/Material.hpp"
+#include "Voxymore/Renderer/Shader.hpp"
+#include "Voxymore/Renderer/UniformBuffer.hpp"
+#include "Voxymore/Renderer/VertexArray.hpp"
 
 #define EXTENSION_GLTF ".gltf"
 #define EXTENSION_GLB ".glb"
