+++ Core/include/Voxymore/RigidbodiesPhysics/Collisions/RigidbodyContactResolver.hpp
@@ -5,8 +5,8 @@
 #pragma once
 
 #include "RigidbodyContact.hpp"
-#include "Voxymore/Core/Math.hpp"
 #include "Voxymore/Core/TimeStep.hpp"
+#include "Voxymore/Math/Math.hpp"
 namespace Voxymore
 {
 	namespace Core
