+++ Core/include/Voxymore/RigidbodiesPhysics/Components/RigidbodySpringComponent.hpp
@@ -4,10 +4,10 @@
 
 #pragma once
 
-#include "Voxymore/Core/Core.hpp"
-#include "Voxymore/Core/Math.hpp"
 #include "Voxymore/Components/Components.hpp"
 #include "Voxymore/Components/CustomComponent.hpp"
+#include "Voxymore/Core/Core.hpp"
+#include "Voxymore/Math/Math.hpp"
 #include "Voxymore/Scene/Entity.hpp"
 
 namespace Voxymore::Core
